% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCryptoCoinsGUI.R
\name{RCryptoCoinsGUI}
\alias{RCryptoCoinsGUI}
\title{User-friendly Interface using RCrypto and quantmod to see the current situation of cryptocurrents}
\usage{
RCryptoCoinsGUI()
}
\description{
User-friendly Interface using RCrypto and quantmod to see the current situation of cryptocurrents
}
\examples{
if(interactive()){
RCryptoCoinsGUI()

}
}
